% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_post.R
\name{new_post}
\alias{new_post}
\title{Create a directory with date_title and index.qmd}
\usage{
new_post(title, author = "default")
}
\arguments{
\item{title}{a string, space will be converted to "_"}
}
\description{
Assumption run at the root of blog
}
\examples{
\dontrun{
new_post("Boule et Bill")
}
}
