% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pot.pourri-package.R
\docType{package}
\name{pot.pourri-package}
\alias{pot.pourri}
\alias{pot.pourri-package}
\title{pot.pourri: clean namespace for mess}
\description{
store some functions that I may run more than once.
}
\author{
\strong{Maintainer}: Olivier Ler \email{leroy.oli@proton.me}

}
\keyword{internal}
