% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_libpath.R
\name{set_proj_libpath}
\alias{set_proj_libpath}
\title{Create a specific directory and set libpath here}
\usage{
set_proj_libpath(my_lib)
}
\arguments{
\item{my_lib}{name of the dir were we want to store pkgs}
}
\description{
Assumption run at the root of the project and ask the user.
Inspired from awong234
}
